requirements.txt
streamlit
numpy
librosa
soundfile
pycryptodomex
reedsolo